simplify's tests separated and doc/src changed
// edited by skirpichev
