+++ docs/modules/simplify/simplify.rst
@@ -7,54 +7,17 @@ simplify
 --------
 .. autofunction:: simplify
 
-collect
--------
-.. autofunction:: collect
-
-.. autofunction:: rcollect
-
 separatevars
 ------------
-
 .. autofunction:: separatevars
 
 nthroot
 -------
 .. autofunction:: nthroot
 
-
-rad_rationalize
----------------
-.. autofunction:: rad_rationalize
-
-
-radsimp
--------
-.. autofunction:: radsimp
-
-ratsimp
--------
-.. autofunction:: ratsimp
-
-fraction
---------
-.. autofunction:: fraction
-
-trigsimp
---------
-.. autofunction:: trigsimp
-
 besselsimp
 ----------
 .. autofunction:: besselsimp
-
-powsimp
--------
-.. autofunction:: powsimp
-
-combsimp
---------
-.. autofunction:: combsimp
 
 hypersimp
 ---------
@@ -68,26 +31,79 @@ nsimplify
 ---------
 .. autofunction:: nsimplify
 
-collect_sqrt
-------------
-.. autofunction:: collect_sqrt
-
-collect_const
--------------
-.. autofunction:: collect_const
-
 posify
 ------
 .. autofunction:: posify
-
-powdenest
----------
-.. autofunction:: powdenest
 
 logcombine
 ----------
 .. autofunction:: logcombine
 
+Radsimp
+-------
+.. module:: sympy.simplify.radsimp
+
+radsimp
+^^^^^^^
+.. autofunction:: radsimp
+
+rad_rationalize
+^^^^^^^^^^^^^^^
+.. autofunction:: rad_rationalize
+
+collect
+^^^^^^^
+.. autofunction:: collect
+
+.. autofunction:: rcollect
+
+collect_sqrt
+^^^^^^^^^^^^
+.. autofunction:: collect_sqrt
+
+collect_const
+^^^^^^^^^^^^^
+.. autofunction:: collect_const
+
+fraction
+^^^^^^^^
+.. autofunction:: fraction
+
+Ratsimp
+-------
+.. module:: sympy.simplify.ratsimp
+
+ratsimp
+^^^^^^^
+.. autofunction:: ratsimp
+
+Trigonometric simplification
+----------------------------
+.. module:: sympy.simplify.trigsimp
+
+trigsimp
+^^^^^^^^
+.. autofunction:: trigsimp
+
+Power simplification
+--------------------
+.. module:: sympy.simplify.powsimp
+
+powsimp
+^^^^^^^
+.. autofunction:: powsimp
+
+powdenest
+^^^^^^^^^
+.. autofunction:: powdenest
+
+Combinatorial simplification
+----------------------------
+.. module:: sympy.simplify.combsimp
+
+combsimp
+^^^^^^^^
+.. autofunction:: combsimp
 
 Square Root Denest
 ------------------
