+++ sympy/simplify/__init__.py
@@ -5,7 +5,7 @@ the expression E**(pi*I) will be converted into -1
 the expression (x+x)**2 will be converted into 4*x**2
 """
 from .simplify import (simplify, hypersimp, hypersimilar,
-    logcombine, separatevars, posify,
+    logcombine, separatevars, posify, besselsimp,
     signsimp, bottom_up, nsimplify)
 
 from .fu import FU, fu
@@ -23,8 +23,6 @@ from .hyperexpand import hyperexpand
 from .radsimp import collect, rcollect, radsimp, collect_const, fraction, numer, denom
 
 from .trigsimp import trigsimp, exptrigsimp
-
-from .besselsimp import besselsimp
 
 from .powsimp import powsimp, powdenest
