+++ sympy/simplify/tests/test_simplify.py
@@ -1,418 +1,22 @@
 from sympy import (
-    Abs, acos, Add, atan, Basic, besselsimp, binomial, collect,
-    collect_const, combsimp, cos, cosh, cot, coth, count_ops,
-    Derivative, diff, Dummy, E, Eq, erf, exp, exp_polar, expand,
-    expand_multinomial, exptrigsimp, factor, factorial, FallingFactorial,
-    Float, fraction, Function, gamma, GoldenRatio, hyper,
-    hypersimp, I, Integral, integrate, log, logcombine, Matrix, MatrixSymbol,
-    Mul, nsimplify, O, oo, pi, Piecewise,
-    posify, powdenest, powsimp, rad, radsimp, Rational, ratsimp,
-    ratsimpmodprime, rcollect, RisingFactorial, root, S, separatevars,
-    signsimp, simplify, sin, sinh, solve, sqrt, Subs, Symbol, symbols,
-    sympify, tan, tanh, trigsimp, Wild, zoo, Sum, Lt)
-from sympy.core.mul import _keep_coeff, _unevaluated_Mul as umul
-from sympy.simplify.simplify import (
-    collect_sqrt, _unevaluated_Add, nthroot)
-from sympy.simplify.radsimp import fraction_expand
+    Abs, acos, Add, atan, Basic, binomial, besselsimp, collect,cos, cosh, cot,
+    coth, count_ops, Derivative, diff, E, Eq, erf, exp, exp_polar, expand,
+    expand_multinomial, factor, factorial, Float, fraction, Function,
+    gamma, GoldenRatio, hyper, hypersimp, I, Integral, integrate, log,
+    logcombine, Matrix, MatrixSymbol, Mul, nsimplify, O, oo, pi, Piecewise,
+    posify, rad, Rational, root, S, separatevars, signsimp, simplify,
+    sin, sinh, solve, sqrt, Symbol, symbols, sympify, tan, tanh, zoo, Sum, Lt)
+from sympy.core.mul import _keep_coeff
+from sympy.simplify.simplify import nthroot
 from sympy.utilities.pytest import XFAIL, slow
 from sympy.core.compatibility import range
 
 from sympy.abc import x, y, z, t, a, b, c, d, e, f, g, h, i, k
 
 
-def test_ratsimp():
-    f, g = 1/x + 1/y, (x + y)/(x*y)
-
-    assert f != g and ratsimp(f) == g
-
-    f, g = 1/(1 + 1/x), 1 - 1/(x + 1)
-
-    assert f != g and ratsimp(f) == g
-
-    f, g = x/(x + y) + y/(x + y), 1
-
-    assert f != g and ratsimp(f) == g
-
-    f, g = -x - y - y**2/(x + y) + x**2/(x + y), -2*y
-
-    assert f != g and ratsimp(f) == g
-
-    f = (a*c*x*y + a*c*z - b*d*x*y - b*d*z - b*t*x*y - b*t*x - b*t*z +
-         e*x)/(x*y + z)
-    G = [a*c - b*d - b*t + (-b*t*x + e*x)/(x*y + z),
-         a*c - b*d - b*t - ( b*t*x - e*x)/(x*y + z)]
-
-    assert f != g and ratsimp(f) in G
-
-    A = sqrt(pi)
-
-    B = log(erf(x) - 1)
-    C = log(erf(x) + 1)
-
-    D = 8 - 8*erf(x)
-
-    f = A*B/D - A*C/D + A*C*erf(x)/D - A*B*erf(x)/D + 2*A/D
-
-    assert ratsimp(f) == A*B/8 - A*C/8 - A/(4*erf(x) - 4)
-
-
-def test_ratsimpmodprime():
-    a = y**5 + x + y
-    b = x - y
-    F = [x*y**5 - x - y]
-    assert ratsimpmodprime(a/b, F, x, y, order='lex') == \
-        (x**2 + x*y + x + y) / (x**2 - x*y)
-
-    a = x + y**2 - 2
-    b = x + y**2 - y - 1
-    F = [x*y - 1]
-    assert ratsimpmodprime(a/b, F, x, y, order='lex') == \
-        (1 + y - x)/(y - x)
-
-    a = 5*x**3 + 21*x**2 + 4*x*y + 23*x + 12*y + 15
-    b = 7*x**3 - y*x**2 + 31*x**2 + 2*x*y + 15*y + 37*x + 21
-    F = [x**2 + y**2 - 1]
-    assert ratsimpmodprime(a/b, F, x, y, order='lex') == \
-        (1 + 5*y - 5*x)/(8*y - 6*x)
-
-    a = x*y - x - 2*y + 4
-    b = x + y**2 - 2*y
-    F = [x - 2, y - 3]
-    assert ratsimpmodprime(a/b, F, x, y, order='lex') == \
-        Rational(2, 5)
-
-    # Test a bug where denominators would be dropped
-    assert ratsimpmodprime(x, [y - 2*x], order='lex') == \
-        y/2
-
-
-def test_trigsimp1():
-    x, y = symbols('x,y')
-
-    assert trigsimp(1 - sin(x)**2) == cos(x)**2
-    assert trigsimp(1 - cos(x)**2) == sin(x)**2
-    assert trigsimp(sin(x)**2 + cos(x)**2) == 1
-    assert trigsimp(1 + tan(x)**2) == 1/cos(x)**2
-    assert trigsimp(1/cos(x)**2 - 1) == tan(x)**2
-    assert trigsimp(1/cos(x)**2 - tan(x)**2) == 1
-    assert trigsimp(1 + cot(x)**2) == 1/sin(x)**2
-    assert trigsimp(1/sin(x)**2 - 1) == 1/tan(x)**2
-    assert trigsimp(1/sin(x)**2 - cot(x)**2) == 1
-
-    assert trigsimp(5*cos(x)**2 + 5*sin(x)**2) == 5
-    assert trigsimp(5*cos(x/2)**2 + 2*sin(x/2)**2) == 3*cos(x)/2 + S(7)/2
-
-    assert trigsimp(sin(x)/cos(x)) == tan(x)
-    assert trigsimp(2*tan(x)*cos(x)) == 2*sin(x)
-    assert trigsimp(cot(x)**3*sin(x)**3) == cos(x)**3
-    assert trigsimp(y*tan(x)**2/sin(x)**2) == y/cos(x)**2
-    assert trigsimp(cot(x)/cos(x)) == 1/sin(x)
-
-    assert trigsimp(sin(x + y) + sin(x - y)) == 2*sin(x)*cos(y)
-    assert trigsimp(sin(x + y) - sin(x - y)) == 2*sin(y)*cos(x)
-    assert trigsimp(cos(x + y) + cos(x - y)) == 2*cos(x)*cos(y)
-    assert trigsimp(cos(x + y) - cos(x - y)) == -2*sin(x)*sin(y)
-    assert trigsimp(tan(x + y) - tan(x)/(1 - tan(x)*tan(y))) == \
-        sin(y)/(-sin(y)*tan(x) + cos(y))  # -tan(y)/(tan(x)*tan(y) - 1)
-
-    assert trigsimp(sinh(x + y) + sinh(x - y)) == 2*sinh(x)*cosh(y)
-    assert trigsimp(sinh(x + y) - sinh(x - y)) == 2*sinh(y)*cosh(x)
-    assert trigsimp(cosh(x + y) + cosh(x - y)) == 2*cosh(x)*cosh(y)
-    assert trigsimp(cosh(x + y) - cosh(x - y)) == 2*sinh(x)*sinh(y)
-    assert trigsimp(tanh(x + y) - tanh(x)/(1 + tanh(x)*tanh(y))) == \
-        sinh(y)/(sinh(y)*tanh(x) + cosh(y))
-
-    assert trigsimp(cos(0.12345)**2 + sin(0.12345)**2) == 1
-    e = 2*sin(x)**2 + 2*cos(x)**2
-    assert trigsimp(log(e)) == log(2)
-
-
-def test_trigsimp1a():
-    assert trigsimp(sin(2)**2*cos(3)*exp(2)/cos(2)**2) == tan(2)**2*cos(3)*exp(2)
-    assert trigsimp(tan(2)**2*cos(3)*exp(2)*cos(2)**2) == sin(2)**2*cos(3)*exp(2)
-    assert trigsimp(cot(2)*cos(3)*exp(2)*sin(2)) == cos(3)*exp(2)*cos(2)
-    assert trigsimp(tan(2)*cos(3)*exp(2)/sin(2)) == cos(3)*exp(2)/cos(2)
-    assert trigsimp(cot(2)*cos(3)*exp(2)/cos(2)) == cos(3)*exp(2)/sin(2)
-    assert trigsimp(cot(2)*cos(3)*exp(2)*tan(2)) == cos(3)*exp(2)
-    assert trigsimp(sinh(2)*cos(3)*exp(2)/cosh(2)) == tanh(2)*cos(3)*exp(2)
-    assert trigsimp(tanh(2)*cos(3)*exp(2)*cosh(2)) == sinh(2)*cos(3)*exp(2)
-    assert trigsimp(coth(2)*cos(3)*exp(2)*sinh(2)) == cosh(2)*cos(3)*exp(2)
-    assert trigsimp(tanh(2)*cos(3)*exp(2)/sinh(2)) == cos(3)*exp(2)/cosh(2)
-    assert trigsimp(coth(2)*cos(3)*exp(2)/cosh(2)) == cos(3)*exp(2)/sinh(2)
-    assert trigsimp(coth(2)*cos(3)*exp(2)*tanh(2)) == cos(3)*exp(2)
-
-
-def test_trigsimp2():
-    x, y = symbols('x,y')
-    assert trigsimp(cos(x)**2*sin(y)**2 + cos(x)**2*cos(y)**2 + sin(x)**2,
-            recursive=True) == 1
-    assert trigsimp(sin(x)**2*sin(y)**2 + sin(x)**2*cos(y)**2 + cos(x)**2,
-            recursive=True) == 1
-    assert trigsimp(
-        Subs(x, x, sin(y)**2 + cos(y)**2)) == Subs(x, x, 1)
-
-
-def test_issue_4373():
-    x = Symbol("x")
-    assert abs(trigsimp(2.0*sin(x)**2 + 2.0*cos(x)**2) - 2.0) < 1e-10
-
-
-def test_trigsimp3():
-    x, y = symbols('x,y')
-    assert trigsimp(sin(x)/cos(x)) == tan(x)
-    assert trigsimp(sin(x)**2/cos(x)**2) == tan(x)**2
-    assert trigsimp(sin(x)**3/cos(x)**3) == tan(x)**3
-    assert trigsimp(sin(x)**10/cos(x)**10) == tan(x)**10
-
-    assert trigsimp(cos(x)/sin(x)) == 1/tan(x)
-    assert trigsimp(cos(x)**2/sin(x)**2) == 1/tan(x)**2
-    assert trigsimp(cos(x)**10/sin(x)**10) == 1/tan(x)**10
-
-    assert trigsimp(tan(x)) == trigsimp(sin(x)/cos(x))
-
-
-def test_issue_4661():
-    a, x, y = symbols('a x y')
-    eq = -4*sin(x)**4 + 4*cos(x)**4 - 8*cos(x)**2
-    assert trigsimp(eq) == -4
-    n = sin(x)**6 + 4*sin(x)**4*cos(x)**2 + 5*sin(x)**2*cos(x)**4 + 2*cos(x)**6
-    d = -sin(x)**2 - 2*cos(x)**2
-    assert simplify(n/d) == -1
-    assert trigsimp(-2*cos(x)**2 + cos(x)**4 - sin(x)**4) == -1
-    eq = (- sin(x)**3/4)*cos(x) + (cos(x)**3/4)*sin(x) - sin(2*x)*cos(2*x)/8
-    assert trigsimp(eq) == 0
-
-
-def test_issue_4494():
-    a, b = symbols('a b')
-    eq = sin(a)**2*sin(b)**2 + cos(a)**2*cos(b)**2*tan(a)**2 + cos(a)**2
-    assert trigsimp(eq) == 1
-
-
-def test_issue_5948():
-    a, x, y = symbols('a x y')
-    assert trigsimp(diff(integrate(cos(x)/sin(x)**7, x), x)) == \
-           cos(x)/sin(x)**7
-
-
-def test_issue_4775():
-    a, x, y = symbols('a x y')
-    assert trigsimp(sin(x)*cos(y)+cos(x)*sin(y)) == sin(x + y)
-    assert trigsimp(sin(x)*cos(y)+cos(x)*sin(y)+3) == sin(x + y) + 3
-
-
-def test_issue_4280():
-    a, x, y = symbols('a x y')
-    assert trigsimp(cos(x)**2 + cos(y)**2*sin(x)**2 + sin(y)**2*sin(x)**2) == 1
-    assert trigsimp(a**2*sin(x)**2 + a**2*cos(y)**2*cos(x)**2 + a**2*cos(x)**2*sin(y)**2) == a**2
-    assert trigsimp(a**2*cos(y)**2*sin(x)**2 + a**2*sin(y)**2*sin(x)**2) == a**2*sin(x)**2
-
-
-def test_issue_3210():
-    eqs = (sin(2)*cos(3) + sin(3)*cos(2),
-        -sin(2)*sin(3) + cos(2)*cos(3),
-        sin(2)*cos(3) - sin(3)*cos(2),
-        sin(2)*sin(3) + cos(2)*cos(3),
-        sin(2)*sin(3) + cos(2)*cos(3) + cos(2),
-        sinh(2)*cosh(3) + sinh(3)*cosh(2),
-        sinh(2)*sinh(3) + cosh(2)*cosh(3),
-        )
-    assert [trigsimp(e) for e in eqs] == [
-        sin(5),
-        cos(5),
-        -sin(1),
-        cos(1),
-        cos(1) + cos(2),
-        sinh(5),
-        cosh(5),
-    ]
-
-
 def test_issue_7263():
     assert abs((simplify(30.8**2 - 82.5**2 * sin(rad(11.6))**2)).evalf() -
             673.447451402970) < 1e-12
-
-
-def test_trigsimp_issues():
-    a, x, y = symbols('a x y')
-
-    # issue 4625 - factor_terms works, too
-    assert trigsimp(sin(x)**3 + cos(x)**2*sin(x)) == sin(x)
-
-    # issue 5948
-    assert trigsimp(diff(integrate(cos(x)/sin(x)**3, x), x)) == \
-        cos(x)/sin(x)**3
-    assert trigsimp(diff(integrate(sin(x)/cos(x)**3, x), x)) == \
-        sin(x)/cos(x)**3
-
-    # check integer exponents
-    e = sin(x)**y/cos(x)**y
-    assert trigsimp(e) == e
-    assert trigsimp(e.subs(y, 2)) == tan(x)**2
-    assert trigsimp(e.subs(x, 1)) == tan(1)**y
-
-    # check for multiple patterns
-    assert (cos(x)**2/sin(x)**2*cos(y)**2/sin(y)**2).trigsimp() == \
-        1/tan(x)**2/tan(y)**2
-    assert trigsimp(cos(x)/sin(x)*cos(x+y)/sin(x+y)) == \
-        1/(tan(x)*tan(x + y))
-
-    eq = cos(2)*(cos(3) + 1)**2/(cos(3) - 1)**2
-    assert trigsimp(eq) == eq.factor()  # factor makes denom (-1 + cos(3))**2
-    assert trigsimp(cos(2)*(cos(3) + 1)**2*(cos(3) - 1)**2) == \
-        cos(2)*sin(3)**4
-
-    # issue 6789; this generates an expression that formerly caused
-    # trigsimp to hang
-    assert cot(x).equals(tan(x)) is False
-
-    # nan or the unchanged expression is ok, but not sin(1)
-    z = cos(x)**2 + sin(x)**2 - 1
-    z1 = tan(x)**2 - 1/cot(x)**2
-    n = (1 + z1/z)
-    assert trigsimp(sin(n)) != sin(1)
-    eq = x*(n - 1) - x*n
-    assert trigsimp(eq) is S.NaN
-    assert trigsimp(eq, recursive=True) is S.NaN
-    assert trigsimp(1).is_Integer
-
-    assert trigsimp(-sin(x)**4 - 2*sin(x)**2*cos(x)**2 - cos(x)**4) == -1
-
-
-def test_trigsimp_issue_2515():
-    x = Symbol('x')
-    assert trigsimp(x*cos(x)*tan(x)) == x*sin(x)
-    assert trigsimp(-sin(x) + cos(x)*tan(x)) == 0
-
-
-def test_trigsimp_issue_3826():
-    assert trigsimp(tan(2*x).expand(trig=True)) == tan(2*x)
-
-
-def test_trigsimp_issue_4032():
-    n = Symbol('n', integer=True, positive=True)
-    assert trigsimp(2**(n/2)*cos(pi*n/4)/2 + 2**(n - 1)/2) == \
-        2**(n/2)*cos(pi*n/4)/2 + 2**n/4
-
-
-def test_trigsimp_issue_7761():
-    assert trigsimp(cosh(pi/4)) == cosh(pi/4)
-
-
-def test_trigsimp_noncommutative():
-    x, y = symbols('x,y')
-    A, B = symbols('A,B', commutative=False)
-
-    assert trigsimp(A - A*sin(x)**2) == A*cos(x)**2
-    assert trigsimp(A - A*cos(x)**2) == A*sin(x)**2
-    assert trigsimp(A*sin(x)**2 + A*cos(x)**2) == A
-    assert trigsimp(A + A*tan(x)**2) == A/cos(x)**2
-    assert trigsimp(A/cos(x)**2 - A) == A*tan(x)**2
-    assert trigsimp(A/cos(x)**2 - A*tan(x)**2) == A
-    assert trigsimp(A + A*cot(x)**2) == A/sin(x)**2
-    assert trigsimp(A/sin(x)**2 - A) == A/tan(x)**2
-    assert trigsimp(A/sin(x)**2 - A*cot(x)**2) == A
-
-    assert trigsimp(y*A*cos(x)**2 + y*A*sin(x)**2) == y*A
-
-    assert trigsimp(A*sin(x)/cos(x)) == A*tan(x)
-    assert trigsimp(A*tan(x)*cos(x)) == A*sin(x)
-    assert trigsimp(A*cot(x)**3*sin(x)**3) == A*cos(x)**3
-    assert trigsimp(y*A*tan(x)**2/sin(x)**2) == y*A/cos(x)**2
-    assert trigsimp(A*cot(x)/cos(x)) == A/sin(x)
-
-    assert trigsimp(A*sin(x + y) + A*sin(x - y)) == 2*A*sin(x)*cos(y)
-    assert trigsimp(A*sin(x + y) - A*sin(x - y)) == 2*A*sin(y)*cos(x)
-    assert trigsimp(A*cos(x + y) + A*cos(x - y)) == 2*A*cos(x)*cos(y)
-    assert trigsimp(A*cos(x + y) - A*cos(x - y)) == -2*A*sin(x)*sin(y)
-
-    assert trigsimp(A*sinh(x + y) + A*sinh(x - y)) == 2*A*sinh(x)*cosh(y)
-    assert trigsimp(A*sinh(x + y) - A*sinh(x - y)) == 2*A*sinh(y)*cosh(x)
-    assert trigsimp(A*cosh(x + y) + A*cosh(x - y)) == 2*A*cosh(x)*cosh(y)
-    assert trigsimp(A*cosh(x + y) - A*cosh(x - y)) == 2*A*sinh(x)*sinh(y)
-
-    assert trigsimp(A*cos(0.12345)**2 + A*sin(0.12345)**2) == 1.0*A
-
-
-def test_hyperbolic_simp():
-    x, y = symbols('x,y')
-
-    assert trigsimp(sinh(x)**2 + 1) == cosh(x)**2
-    assert trigsimp(cosh(x)**2 - 1) == sinh(x)**2
-    assert trigsimp(cosh(x)**2 - sinh(x)**2) == 1
-    assert trigsimp(1 - tanh(x)**2) == 1/cosh(x)**2
-    assert trigsimp(1 - 1/cosh(x)**2) == tanh(x)**2
-    assert trigsimp(tanh(x)**2 + 1/cosh(x)**2) == 1
-    assert trigsimp(coth(x)**2 - 1) == 1/sinh(x)**2
-    assert trigsimp(1/sinh(x)**2 + 1) == 1/tanh(x)**2
-    assert trigsimp(coth(x)**2 - 1/sinh(x)**2) == 1
-
-    assert trigsimp(5*cosh(x)**2 - 5*sinh(x)**2) == 5
-    assert trigsimp(5*cosh(x/2)**2 - 2*sinh(x/2)**2) == 3*cosh(x)/2 + S(7)/2
-
-    assert trigsimp(sinh(x)/cosh(x)) == tanh(x)
-    assert trigsimp(tanh(x)) == trigsimp(sinh(x)/cosh(x))
-    assert trigsimp(cosh(x)/sinh(x)) == 1/tanh(x)
-    assert trigsimp(2*tanh(x)*cosh(x)) == 2*sinh(x)
-    assert trigsimp(coth(x)**3*sinh(x)**3) == cosh(x)**3
-    assert trigsimp(y*tanh(x)**2/sinh(x)**2) == y/cosh(x)**2
-    assert trigsimp(coth(x)/cosh(x)) == 1/sinh(x)
-
-    e = 2*cosh(x)**2 - 2*sinh(x)**2
-    assert trigsimp(log(e)) == log(2)
-
-    assert trigsimp(cosh(x)**2*cosh(y)**2 - cosh(x)**2*sinh(y)**2 - sinh(x)**2,
-            recursive=True) == 1
-    assert trigsimp(sinh(x)**2*sinh(y)**2 - sinh(x)**2*cosh(y)**2 + cosh(x)**2,
-            recursive=True) == 1
-
-    assert abs(trigsimp(2.0*cosh(x)**2 - 2.0*sinh(x)**2) - 2.0) < 1e-10
-
-    assert trigsimp(sinh(x)**2/cosh(x)**2) == tanh(x)**2
-    assert trigsimp(sinh(x)**3/cosh(x)**3) == tanh(x)**3
-    assert trigsimp(sinh(x)**10/cosh(x)**10) == tanh(x)**10
-    assert trigsimp(cosh(x)**3/sinh(x)**3) == 1/tanh(x)**3
-
-    assert trigsimp(cosh(x)/sinh(x)) == 1/tanh(x)
-    assert trigsimp(cosh(x)**2/sinh(x)**2) == 1/tanh(x)**2
-    assert trigsimp(cosh(x)**10/sinh(x)**10) == 1/tanh(x)**10
-
-    assert trigsimp(x*cosh(x)*tanh(x)) == x*sinh(x)
-    assert trigsimp(-sinh(x) + cosh(x)*tanh(x)) == 0
-
-    assert tan(x) != 1/cot(x)  # cot doesn't auto-simplify
-
-    assert trigsimp(tan(x) - 1/cot(x)) == 0
-    assert trigsimp(3*tanh(x)**7 - 2/coth(x)**7) == tanh(x)**7
-
-
-def test_trigsimp_groebner():
-    from sympy.simplify.trigsimp import trigsimp_groebner
-
-    c = cos(x)
-    s = sin(x)
-    ex = (4*s*c + 12*s + 5*c**3 + 21*c**2 + 23*c + 15)/(
-        -s*c**2 + 2*s*c + 15*s + 7*c**3 + 31*c**2 + 37*c + 21)
-    resnum = (5*s - 5*c + 1)
-    resdenom = (8*s - 6*c)
-    results = [resnum/resdenom, (-resnum)/(-resdenom)]
-    assert trigsimp_groebner(ex) in results
-    assert trigsimp_groebner(s/c, hints=[tan]) == tan(x)
-
-    assert trigsimp((-s + 1)/c + c/(-s + 1),
-                    method='groebner') == 2/c
-    assert trigsimp((-s + 1)/c + c/(-s + 1),
-                    method='groebner', polynomial=True) == 2/c
-
-    # Test quick=False works
-    assert trigsimp_groebner(ex, hints=[2]) in results
-
-    # test "I"
-    assert trigsimp_groebner(sin(I*x)/cos(I*x), hints=[tanh]) == I*tanh(x)
-
-    # test hyperbolic / sums
-    assert trigsimp_groebner((tanh(x)+tanh(y))/(1+tanh(x)*tanh(y)),
-                             hints=[(tanh, x, y)]) == tanh(x + y)
 
 
 @XFAIL
@@ -568,207 +172,6 @@ def test_simplify_fail1():
     assert simplify(e) == 1 / (-2*y)
 
 
-def test_fraction():
-    x, y, z = map(Symbol, 'xyz')
-    A = Symbol('A', commutative=False)
-
-    assert fraction(Rational(1, 2)) == (1, 2)
-
-    assert fraction(x) == (x, 1)
-    assert fraction(1/x) == (1, x)
-    assert fraction(x/y) == (x, y)
-    assert fraction(x/2) == (x, 2)
-
-    assert fraction(x*y/z) == (x*y, z)
-    assert fraction(x/(y*z)) == (x, y*z)
-
-    assert fraction(1/y**2) == (1, y**2)
-    assert fraction(x/y**2) == (x, y**2)
-
-    assert fraction((x**2 + 1)/y) == (x**2 + 1, y)
-    assert fraction(x*(y + 1)/y**7) == (x*(y + 1), y**7)
-
-    assert fraction(exp(-x), exact=True) == (exp(-x), 1)
-
-    assert fraction(x*A/y) == (x*A, y)
-    assert fraction(x*A**-1/y) == (x*A**-1, y)
-
-    n = symbols('n', negative=True)
-    assert fraction(exp(n)) == (1, exp(-n))
-    assert fraction(exp(-n)) == (exp(-n), 1)
-
-
-def test_powsimp():
-    x, y, z, n = symbols('x,y,z,n')
-    f = Function('f')
-    assert powsimp( 4**x * 2**(-x) * 2**(-x) ) == 1
-    assert powsimp( (-4)**x * (-2)**(-x) * 2**(-x) ) == 1
-
-    assert powsimp(
-        f(4**x * 2**(-x) * 2**(-x)) ) == f(4**x * 2**(-x) * 2**(-x))
-    assert powsimp( f(4**x * 2**(-x) * 2**(-x)), deep=True ) == f(1)
-    assert exp(x)*exp(y) == exp(x)*exp(y)
-    assert powsimp(exp(x)*exp(y)) == exp(x + y)
-    assert powsimp(exp(x)*exp(y)*2**x*2**y) == (2*E)**(x + y)
-    assert powsimp(exp(x)*exp(y)*2**x*2**y, combine='exp') == \
-        exp(x + y)*2**(x + y)
-    assert powsimp(exp(x)*exp(y)*exp(2)*sin(x) + sin(y) + 2**x*2**y) == \
-        exp(2 + x + y)*sin(x) + sin(y) + 2**(x + y)
-    assert powsimp(sin(exp(x)*exp(y))) == sin(exp(x)*exp(y))
-    assert powsimp(sin(exp(x)*exp(y)), deep=True) == sin(exp(x + y))
-    assert powsimp(x**2*x**y) == x**(2 + y)
-    # This should remain factored, because 'exp' with deep=True is supposed
-    # to act like old automatic exponent combining.
-    assert powsimp((1 + E*exp(E))*exp(-E), combine='exp', deep=True) == \
-        (1 + exp(1 + E))*exp(-E)
-    assert powsimp((1 + E*exp(E))*exp(-E), deep=True) == \
-        (1 + exp(1 + E))*exp(-E)
-    assert powsimp((1 + E*exp(E))*exp(-E)) == (1 + exp(1 + E))*exp(-E)
-    assert powsimp((1 + E*exp(E))*exp(-E), combine='exp') == \
-        (1 + exp(1 + E))*exp(-E)
-    assert powsimp((1 + E*exp(E))*exp(-E), combine='base') == \
-        (1 + E*exp(E))*exp(-E)
-    x, y = symbols('x,y', nonnegative=True)
-    n = Symbol('n', extended_real=True)
-    assert powsimp(y**n * (y/x)**(-n)) == x**n
-    assert powsimp(x**(x**(x*y)*y**(x*y))*y**(x**(x*y)*y**(x*y)), deep=True) \
-        == (x*y)**(x*y)**(x*y)
-    assert powsimp(2**(2**(2*x)*x), deep=False) == 2**(2**(2*x)*x)
-    assert powsimp(2**(2**(2*x)*x), deep=True) == 2**(x*4**x)
-    assert powsimp(
-        exp(-x + exp(-x)*exp(-x*log(x))), deep=False, combine='exp') == \
-        exp(-x + exp(-x)*exp(-x*log(x)))
-    assert powsimp(
-        exp(-x + exp(-x)*exp(-x*log(x))), deep=False, combine='exp') == \
-        exp(-x + exp(-x)*exp(-x*log(x)))
-    assert powsimp((x + y)/(3*z), deep=False, combine='exp') == (x + y)/(3*z)
-    assert powsimp((x/3 + y/3)/z, deep=True, combine='exp') == (x/3 + y/3)/z
-    assert powsimp(exp(x)/(1 + exp(x)*exp(y)), deep=True) == \
-        exp(x)/(1 + exp(x + y))
-    assert powsimp(x*y**(z**x*z**y), deep=True) == x*y**(z**(x + y))
-    assert powsimp((z**x*z**y)**x, deep=True) == (z**(x + y))**x
-    assert powsimp(x*(z**x*z**y)**x, deep=True) == x*(z**(x + y))**x
-    p = symbols('p', positive=True)
-    assert powsimp((1/x)**log(2)/x) == (1/x)**(1 + log(2))
-    assert powsimp((1/p)**log(2)/p) == p**(-1 - log(2))
-
-    # coefficient of exponent can only be simplified for positive bases
-    assert powsimp(2**(2*x)) == 4**x
-    assert powsimp((-1)**(2*x)) == (-1)**(2*x)
-    i = symbols('i', integer=True)
-    assert powsimp((-1)**(2*i)) == 1
-    assert powsimp((-1)**(-x)) != (-1)**x  # could be 1/((-1)**x), but is not
-    # force=True overrides assumptions
-    assert powsimp((-1)**(2*x), force=True) == 1
-
-    # rational exponents allow combining of negative terms
-    w, n, m = symbols('w n m', negative=True)
-    e = i/a  # not a rational exponent if `a` is unknown
-    ex = w**e*n**e*m**e
-    assert powsimp(ex) == m**(i/a)*n**(i/a)*w**(i/a)
-    e = i/3
-    ex = w**e*n**e*m**e
-    assert powsimp(ex) == (-1)**i*(-m*n*w)**(i/3)
-    e = (3 + i)/i
-    ex = w**e*n**e*m**e
-    assert powsimp(ex) == (-1)**(3*e)*(-m*n*w)**e
-
-    eq = x**(2*a/3)
-    # eq != (x**a)**(2/3) (try x = -1 and a = 3 to see)
-    assert powsimp(eq).exp == eq.exp == 2*a/3
-    # powdenest goes the other direction
-    assert powsimp(2**(2*x)) == 4**x
-
-    assert powsimp(exp(p/2)) == exp(p/2)
-
-    # issue 6368
-    eq = Mul(*[sqrt(Dummy(imaginary=True)) for i in range(3)])
-    assert powsimp(eq) == eq and eq.is_Mul
-
-    assert all(powsimp(e) == e for e in (sqrt(x**a), sqrt(x**2)))
-
-    # issue 8836
-    assert str( powsimp(exp(I*pi/3)*root(-1,3)) ) == '(-1)**(2/3)'
-
-
-def test_issue_6367():
-    z = -5*sqrt(2)/(2*sqrt(2*sqrt(29) + 29)) + sqrt(-sqrt(29)/29 + S(1)/2)
-    assert Mul(*[powsimp(a) for a in Mul.make_args(z.normal())]) == 0
-    assert powsimp(z.normal()) == 0
-    assert simplify(z) == 0
-    assert powsimp(sqrt(2 + sqrt(3))*sqrt(2 - sqrt(3)) + 1) == 2
-    assert powsimp(z) != 0
-
-
-def test_powsimp_negated_base():
-    assert powsimp((-x + y)/sqrt(x - y)) == -sqrt(x - y)
-    assert powsimp((-x + y)*(-z + y)/sqrt(x - y)/sqrt(z - y)) == sqrt(x - y)*sqrt(z - y)
-    p = symbols('p', positive=True)
-    assert powsimp((-p)**a/p**a) == (-1)**a
-    n = symbols('n', negative=True)
-    assert powsimp((-n)**a/n**a) == (-1)**a
-    # if x is 0 then the lhs is 0**a*oo**a which is not (-1)**a
-    assert powsimp((-x)**a/x**a) != (-1)**a
-
-
-def test_issue_6440():
-    assert powsimp(16*2**a*8**b) == 2**(a + 3*b + 4)
-
-
-def test_powsimp_polar():
-    from sympy import polar_lift, exp_polar
-    x, y, z = symbols('x y z')
-    p, q, r = symbols('p q r', polar=True)
-
-    assert (polar_lift(-1))**(2*x) == exp_polar(2*pi*I*x)
-    assert powsimp(p**x * q**x) == (p*q)**x
-    assert p**x * (1/p)**x == 1
-    assert (1/p)**x == p**(-x)
-
-    assert exp_polar(x)*exp_polar(y) == exp_polar(x)*exp_polar(y)
-    assert powsimp(exp_polar(x)*exp_polar(y)) == exp_polar(x + y)
-    assert powsimp(exp_polar(x)*exp_polar(y)*p**x*p**y) == \
-        (p*exp_polar(1))**(x + y)
-    assert powsimp(exp_polar(x)*exp_polar(y)*p**x*p**y, combine='exp') == \
-        exp_polar(x + y)*p**(x + y)
-    assert powsimp(
-        exp_polar(x)*exp_polar(y)*exp_polar(2)*sin(x) + sin(y) + p**x*p**y) \
-        == p**(x + y) + sin(x)*exp_polar(2 + x + y) + sin(y)
-    assert powsimp(sin(exp_polar(x)*exp_polar(y))) == \
-        sin(exp_polar(x)*exp_polar(y))
-    assert powsimp(sin(exp_polar(x)*exp_polar(y)), deep=True) == \
-        sin(exp_polar(x + y))
-
-
-def test_powsimp_nc():
-    x, y, z = symbols('x,y,z')
-    A, B, C = symbols('A B C', commutative=False)
-
-    assert powsimp(A**x*A**y, combine='all') == A**(x + y)
-    assert powsimp(A**x*A**y, combine='base') == A**x*A**y
-    assert powsimp(A**x*A**y, combine='exp') == A**(x + y)
-
-    assert powsimp(A**x*B**x, combine='all') == (A*B)**x
-    assert powsimp(A**x*B**x, combine='base') == (A*B)**x
-    assert powsimp(A**x*B**x, combine='exp') == A**x*B**x
-
-    assert powsimp(B**x*A**x, combine='all') == (B*A)**x
-    assert powsimp(B**x*A**x, combine='base') == (B*A)**x
-    assert powsimp(B**x*A**x, combine='exp') == B**x*A**x
-
-    assert powsimp(A**x*A**y*A**z, combine='all') == A**(x + y + z)
-    assert powsimp(A**x*A**y*A**z, combine='base') == A**x*A**y*A**z
-    assert powsimp(A**x*A**y*A**z, combine='exp') == A**(x + y + z)
-
-    assert powsimp(A**x*B**x*C**x, combine='all') == (A*B*C)**x
-    assert powsimp(A**x*B**x*C**x, combine='base') == (A*B*C)**x
-    assert powsimp(A**x*B**x*C**x, combine='exp') == A**x*B**x*C**x
-
-    assert powsimp(B**x*A**x*C**x, combine='all') == (B*A*C)**x
-    assert powsimp(B**x*A**x*C**x, combine='base') == (B*A*C)**x
-    assert powsimp(B**x*A**x*C**x, combine='exp') == B**x*A**x*C**x
-
-
 def test_nthroot():
     assert nthroot(90 + 34*sqrt(7), 3) == sqrt(7) + 3
     q = 1 + sqrt(2) - 2*sqrt(3) + sqrt(6) + sqrt(7)
@@ -794,179 +197,6 @@ def test_nthroot1():
     q = 1 + sqrt(2) + sqrt(3) + S(1)/10**30
     p = expand_multinomial(q**5)
     assert nthroot(p, 5) == q
-
-
-def test_collect_1():
-    """Collect with respect to a Symbol"""
-    x, y, z, n = symbols('x,y,z,n')
-    assert collect( x + y*x, x ) == x * (1 + y)
-    assert collect( x + x**2, x ) == x + x**2
-    assert collect( x**2 + y*x**2, x ) == (x**2)*(1 + y)
-    assert collect( x**2 + y*x, x ) == x*y + x**2
-    assert collect( 2*x**2 + y*x**2 + 3*x*y, [x] ) == x**2*(2 + y) + 3*x*y
-    assert collect( 2*x**2 + y*x**2 + 3*x*y, [y] ) == 2*x**2 + y*(x**2 + 3*x)
-
-    assert collect( ((1 + y + x)**4).expand(), x) == ((1 + y)**4).expand() + \
-        x*(4*(1 + y)**3).expand() + x**2*(6*(1 + y)**2).expand() + \
-        x**3*(4*(1 + y)).expand() + x**4
-    # symbols can be given as any iterable
-    expr = x + y
-    assert collect(expr, expr.free_symbols) == expr
-
-
-def test_collect_2():
-    """Collect with respect to a sum"""
-    a, b, x = symbols('a,b,x')
-    assert collect(a*(cos(x) + sin(x)) + b*(cos(x) + sin(x)),
-        sin(x) + cos(x)) == (a + b)*(cos(x) + sin(x))
-
-
-def test_collect_3():
-    """Collect with respect to a product"""
-    a, b, c = symbols('a,b,c')
-    f = Function('f')
-    x, y, z, n = symbols('x,y,z,n')
-
-    assert collect(-x/8 + x*y, -x) == x*(y - S(1)/8)
-
-    assert collect( 1 + x*(y**2), x*y ) == 1 + x*(y**2)
-    assert collect( x*y + a*x*y, x*y) == x*y*(1 + a)
-    assert collect( 1 + x*y + a*x*y, x*y) == 1 + x*y*(1 + a)
-    assert collect(a*x*f(x) + b*(x*f(x)), x*f(x)) == x*(a + b)*f(x)
-
-    assert collect(a*x*log(x) + b*(x*log(x)), x*log(x)) == x*(a + b)*log(x)
-    assert collect(a*x**2*log(x)**2 + b*(x*log(x))**2, x*log(x)) == \
-        x**2*log(x)**2*(a + b)
-
-    # with respect to a product of three symbols
-    assert collect(y*x*z + a*x*y*z, x*y*z) == (1 + a)*x*y*z
-
-
-def test_collect_4():
-    """Collect with respect to a power"""
-    a, b, c, x = symbols('a,b,c,x')
-
-    assert collect(a*x**c + b*x**c, x**c) == x**c*(a + b)
-    # issue 6096: 2 stays with c (unless c is integer or x is positive0
-    assert collect(a*x**(2*c) + b*x**(2*c), x**c) == x**(2*c)*(a + b)
-
-
-def test_collect_5():
-    """Collect with respect to a tuple"""
-    a, x, y, z, n = symbols('a,x,y,z,n')
-    assert collect(x**2*y**4 + z*(x*y**2)**2 + z + a*z, [x*y**2, z]) in [
-        z*(1 + a + x**2*y**4) + x**2*y**4,
-        z*(1 + a) + x**2*y**4*(1 + z) ]
-    assert collect((1 + (x + y) + (x + y)**2).expand(),
-                   [x, y]) == 1 + y + x*(1 + 2*y) + x**2 + y**2
-
-
-def test_collect_D():
-    D = Derivative
-    f = Function('f')
-    x, a, b = symbols('x,a,b')
-    fx = D(f(x), x)
-    fxx = D(f(x), x, x)
-
-    assert collect(a*fx + b*fx, fx) == (a + b)*fx
-    assert collect(a*D(fx, x) + b*D(fx, x), fx) == (a + b)*D(fx, x)
-    assert collect(a*fxx + b*fxx, fx) == (a + b)*D(fx, x)
-    # issue 4784
-    assert collect(5*f(x) + 3*fx, fx) == 5*f(x) + 3*fx
-    assert collect(f(x) + f(x)*diff(f(x), x) + x*diff(f(x), x)*f(x), f(x).diff(x)) == \
-        (x*f(x) + f(x))*D(f(x), x) + f(x)
-    assert collect(f(x) + f(x)*diff(f(x), x) + x*diff(f(x), x)*f(x), f(x).diff(x), exact=True) == \
-        (x*f(x) + f(x))*D(f(x), x) + f(x)
-    assert collect(1/f(x) + 1/f(x)*diff(f(x), x) + x*diff(f(x), x)/f(x), f(x).diff(x), exact=True) == \
-        (1/f(x) + x/f(x))*D(f(x), x) + 1/f(x)
-
-
-@XFAIL
-def test_collect_issues():
-    D = Derivative
-    f = Function('f')
-    e = (1 + x*D(f(x), x) + D(f(x), x))/f(x)
-    assert collect(e.expand(), f(x).diff(x)) != e
-
-
-def test_collect_D_0():
-    D = Derivative
-    f = Function('f')
-    x, a, b = symbols('x,a,b')
-    fxx = D(f(x), x, x)
-
-    # collect does not distinguish nested derivatives, so it returns
-    #                                           -- (a + b)*D(D(f, x), x)
-    assert collect(a*fxx + b*fxx, fxx) == (a + b)*fxx
-
-
-def test_collect_Wild():
-    """Collect with respect to functions with Wild argument"""
-    a, b, x, y = symbols('a b x y')
-    f = Function('f')
-    w1 = Wild('.1')
-    w2 = Wild('.2')
-    assert collect(f(x) + a*f(x), f(w1)) == (1 + a)*f(x)
-    assert collect(f(x, y) + a*f(x, y), f(w1)) == f(x, y) + a*f(x, y)
-    assert collect(f(x, y) + a*f(x, y), f(w1, w2)) == (1 + a)*f(x, y)
-    assert collect(f(x, y) + a*f(x, y), f(w1, w1)) == f(x, y) + a*f(x, y)
-    assert collect(f(x, x) + a*f(x, x), f(w1, w1)) == (1 + a)*f(x, x)
-    assert collect(a*(x + 1)**y + (x + 1)**y, w1**y) == (1 + a)*(x + 1)**y
-    assert collect(a*(x + 1)**y + (x + 1)**y, w1**b) == \
-        a*(x + 1)**y + (x + 1)**y
-    assert collect(a*(x + 1)**y + (x + 1)**y, (x + 1)**w2) == \
-        (1 + a)*(x + 1)**y
-    assert collect(a*(x + 1)**y + (x + 1)**y, w1**w2) == (1 + a)*(x + 1)**y
-
-
-def test_collect_func():
-    f = ((x + a + 1)**3).expand()
-
-    assert collect(f, x) == a**3 + 3*a**2 + 3*a + x**3 + x**2*(3*a + 3) + \
-        x*(3*a**2 + 6*a + 3) + 1
-    assert collect(f, x, factor) == x**3 + 3*x**2*(a + 1) + 3*x*(a + 1)**2 + \
-        (a + 1)**3
-
-    assert collect(f, x, evaluate=False) == {
-        S.One: a**3 + 3*a**2 + 3*a + 1,
-        x: 3*a**2 + 6*a + 3, x**2: 3*a + 3,
-        x**3: 1
-    }
-
-
-@XFAIL
-def test_collect_func_xfail():
-    # XXX: this test will pass when automatic constant distribution is removed (issue 4596)
-    assert collect(f, x, factor, evaluate=False) == {S.One: (a + 1)**3,
-                   x: 3*(a + 1)**2, x**2: 3*(a + 1), x**3: 1}
-
-
-def test_collect_order():
-    a, b, x, t = symbols('a,b,x,t')
-
-    assert collect(t + t*x + t*x**2 + O(x**3), t) == t*(1 + x + x**2 + O(x**3))
-    assert collect(t + t*x + x**2 + O(x**3), t) == \
-        t*(1 + x + O(x**3)) + x**2 + O(x**3)
-
-    f = a*x + b*x + c*x**2 + d*x**2 + O(x**3)
-    g = x*(a + b) + x**2*(c + d) + O(x**3)
-
-    assert collect(f, x) == g
-    assert collect(f, x, distribute_order_term=False) == g
-
-    f = sin(a + b).series(b, 0, 10)
-
-    assert collect(f, [sin(a), cos(a)]) == \
-        sin(a)*cos(b).series(b, 0, 10) + cos(a)*sin(b).series(b, 0, 10)
-    assert collect(f, [sin(a), cos(a)], distribute_order_term=False) == \
-        sin(a)*cos(b).series(b, 0, 10).removeO() + \
-        cos(a)*sin(b).series(b, 0, 10).removeO() + O(b**10)
-
-
-def test_rcollect():
-    assert rcollect((x**2*y + x*y + x + y)/(x + y), y) == \
-        (x + y*(1 + x + x**2))/(x + y)
-    assert rcollect(sqrt(-((x + 1)*(y + 1))), z) == sqrt(-((x + 1)*(y + 1)))
 
 
 def test_separatevars():
@@ -1215,68 +445,6 @@ def test_posify():
         'Sum(_x**(-n), (n, 1, 3))'
 
 
-def test_powdenest():
-    from sympy import powdenest
-    from sympy.abc import x, y, z, a, b
-    p, q = symbols('p q', positive=True)
-    i, j = symbols('i,j', integer=True)
-
-    assert powdenest(x) == x
-    assert powdenest(x + 2*(x**(2*a/3))**(3*x)) == (x + 2*(x**(2*a/3))**(3*x))
-    assert powdenest((exp(2*a/3))**(3*x))  # -X-> (exp(a/3))**(6*x)
-    assert powdenest((x**(2*a/3))**(3*x)) == ((x**(2*a/3))**(3*x))
-    assert powdenest(exp(3*x*log(2))) == 2**(3*x)
-    assert powdenest(sqrt(p**2)) == p
-    i, j = symbols('i,j', integer=True)
-    eq = p**(2*i)*q**(4*i)
-    assert powdenest(eq) == (p*q**2)**(2*i)
-    # -X-> (x**x)**i*(x**x)**j == x**(x*(i + j))
-    assert powdenest((x**x)**(i + j))
-    assert powdenest(exp(3*y*log(x))) == x**(3*y)
-    assert powdenest(exp(y*(log(a) + log(b)))) == (a*b)**y
-    assert powdenest(exp(3*(log(a) + log(b)))) == a**3*b**3
-    assert powdenest(((x**(2*i))**(3*y))**x) == ((x**(2*i))**(3*y))**x
-    assert powdenest(((x**(2*i))**(3*y))**x, force=True) == x**(6*i*x*y)
-    assert powdenest(((x**(2*a/3))**(3*y/i))**x) == \
-        (((x**(2*a/3))**(3*y/i))**x)
-    assert powdenest((x**(2*i)*y**(4*i))**z, force=True) == (x*y**2)**(2*i*z)
-    assert powdenest((p**(2*i)*q**(4*i))**j) == (p*q**2)**(2*i*j)
-    e = ((p**(2*a))**(3*y))**x
-    assert powdenest(e) == e
-    e = ((x**2*y**4)**a)**(x*y)
-    assert powdenest(e) == e
-    e = (((x**2*y**4)**a)**(x*y))**3
-    assert powdenest(e) == ((x**2*y**4)**a)**(3*x*y)
-    assert powdenest((((x**2*y**4)**a)**(x*y)), force=True) == \
-        (x*y**2)**(2*a*x*y)
-    assert powdenest((((x**2*y**4)**a)**(x*y))**3, force=True) == \
-        (x*y**2)**(6*a*x*y)
-    assert powdenest((x**2*y**6)**i) != (x*y**3)**(2*i)
-    x, y = symbols('x,y', positive=True)
-    assert powdenest((x**2*y**6)**i) == (x*y**3)**(2*i)
-
-    assert powdenest((x**(2*i/3)*y**(i/2))**(2*i)) == (x**(S(4)/3)*y)**(i**2)
-    assert powdenest(sqrt(x**(2*i)*y**(6*i))) == (x*y**3)**i
-
-    assert powdenest(4**x) == 2**(2*x)
-    assert powdenest((4**x)**y) == 2**(2*x*y)
-    assert powdenest(4**x*y) == 2**(2*x)*y
-
-
-def test_powdenest_polar():
-    x, y, z = symbols('x y z', polar=True)
-    a, b, c = symbols('a b c')
-    assert powdenest((x*y*z)**a) == x**a*y**a*z**a
-    assert powdenest((x**a*y**b)**c) == x**(a*c)*y**(b*c)
-    assert powdenest(((x**a)**b*y**c)**c) == x**(a*b*c)*y**(c**2)
-
-
-def test_issue_5805():
-    arg = ((gamma(x)*hyper((), (), x))*pi)**2
-    assert powdenest(arg) == (pi*gamma(x)*hyper((), (), x))**2
-    assert arg.is_positive is None
-
-
 def test_issue_4194():
     # simplify should call cancel
     from sympy.abc import x, y
@@ -1290,98 +458,6 @@ def test_simplify_float_vs_integer():
     # https://github.com/sympy/sympy/issues/4473
     assert simplify(x**2.0 - x**2) == 0
     assert simplify(x**2 - x**2.0) == 0
-
-
-def test_combsimp():
-    from sympy.abc import n, k
-
-    assert combsimp(factorial(n)) == factorial(n)
-    assert combsimp(binomial(n, k)) == binomial(n, k)
-
-    assert combsimp(factorial(n)/factorial(n - 3)) == n*(-1 + n)*(-2 + n)
-    assert combsimp(binomial(n + 1, k + 1)/binomial(n, k)) == (1 + n)/(1 + k)
-
-    assert combsimp(binomial(3*n + 4, n + 1)/binomial(3*n + 1, n)) == \
-        S(3)/2*((3*n + 2)*(3*n + 4)/((n + 1)*(2*n + 3)))
-
-    assert combsimp(factorial(n)**2/factorial(n - 3)) == \
-        factorial(n)*n*(-1 + n)*(-2 + n)
-    assert combsimp(factorial(n)*binomial(n + 1, k + 1)/binomial(n, k)) == \
-        factorial(n)*(1 + n)/(1 + k)
-
-    assert combsimp(binomial(n - 1, k)) == -((-n + k)*binomial(n, k))/n
-
-    assert combsimp(binomial(n + 2, k + S(1)/2)) == 4*((n + 1)*(n + 2) *
-        binomial(n, k + S(1)/2))/((2*k - 2*n - 1)*(2*k - 2*n - 3))
-    assert combsimp(binomial(n + 2, k + 2.0)) == \
-        -((1.0*n + 2.0)*binomial(n + 1.0, k + 2.0))/(k - n)
-
-    # coverage tests
-    assert combsimp(factorial(n*(1 + n) - n**2 - n)) == 1
-    assert combsimp(binomial(n + k - 2, n)) == \
-        k*(k - 1)*binomial(n + k, n)/((n + k)*(n + k - 1))
-    i = Symbol('i', integer=True)
-    e = gamma(i + 3)
-    assert combsimp(e) == e
-    e = gamma(exp(i))
-    assert combsimp(e) == e
-    e = gamma(n + S(1)/3)*gamma(n + S(2)/3)
-    assert combsimp(e) == e
-    assert combsimp(gamma(4*n + S(1)/2)/gamma(2*n - S(3)/4)) == \
-        2**(4*n - S(5)/2)*(8*n - 3)*gamma(2*n + S(3)/4)/sqrt(pi)
-
-    assert combsimp(6*FallingFactorial(-4, n)/factorial(n)) == \
-        (-1)**n*(n + 1)*(n + 2)*(n + 3)
-    assert combsimp(6*FallingFactorial(-4, n - 1)/factorial(n - 1)) == \
-        (-1)**(n - 1)*n*(n + 1)*(n + 2)
-    assert combsimp(6*FallingFactorial(-4, n - 3)/factorial(n - 3)) == \
-        (-1)**(n - 3)*n*(n - 1)*(n - 2)
-    assert combsimp(6*FallingFactorial(-4, -n - 1)/factorial(-n - 1)) == \
-        -(-1)**(-n - 1)*n*(n - 1)*(n - 2)
-
-    assert combsimp(6*RisingFactorial(4, n)/factorial(n)) == \
-        (n + 1)*(n + 2)*(n + 3)
-    assert combsimp(6*RisingFactorial(4, n - 1)/factorial(n - 1)) == \
-        n*(n + 1)*(n + 2)
-    assert combsimp(6*RisingFactorial(4, n - 3)/factorial(n - 3)) == \
-        n*(n - 1)*(n - 2)
-    assert combsimp(6*RisingFactorial(4, -n - 1)/factorial(-n - 1)) == \
-        -n*(n - 1)*(n - 2)
-
-
-def test_issue_5615():
-    aA, Re, a, b, D = symbols('aA Re a b D')
-    e = ((D**3*a + b*aA**3)/Re).expand()
-    assert collect(e, [aA**3/Re, a]) == e
-
-
-def test_issue_5728():
-    b = x*sqrt(y)
-    a = sqrt(b)
-    c = sqrt(sqrt(x)*y)
-    assert powsimp(a*b) == sqrt(b)**3
-    assert powsimp(a*b**2*sqrt(y)) == sqrt(y)*a**5
-    assert powsimp(a*x**2*c**3*y) == c**3*a**5
-    assert powsimp(a*x*c**3*y**2) == c**7*a
-    assert powsimp(x*c**3*y**2) == c**7
-    assert powsimp(x*c**3*y) == x*y*c**3
-    assert powsimp(sqrt(x)*c**3*y) == c**5
-    assert powsimp(sqrt(x)*a**3*sqrt(y)) == sqrt(x)*sqrt(y)*a**3
-    assert powsimp(Mul(sqrt(x)*c**3*sqrt(y), y, evaluate=False)) == \
-        sqrt(x)*sqrt(y)**3*c**3
-    assert powsimp(a**2*a*x**2*y) == a**7
-
-    # symbolic powers work, too
-    b = x**y*y
-    a = b*sqrt(b)
-    assert a.is_Mul is True
-    assert powsimp(a) == sqrt(b)**3
-
-    # as does exp
-    a = x*exp(2*y/3)
-    assert powsimp(a*sqrt(a)) == sqrt(a)**3
-    assert powsimp(a**2*sqrt(a)) == sqrt(a)**5
-    assert powsimp(a**2*sqrt(sqrt(a))) == sqrt(sqrt(a))**9
 
 
 def test_as_content_primitive():
@@ -1423,337 +499,6 @@ def test_as_content_primitive():
            (1, 2**(S(1)/4)*(sqrt(2) + sqrt(3)))
 
 
-def test_radsimp():
-    r2 = sqrt(2)
-    r3 = sqrt(3)
-    r5 = sqrt(5)
-    r7 = sqrt(7)
-    assert fraction(radsimp(1/r2)) == (sqrt(2), 2)
-    assert radsimp(1/(1 + r2)) == \
-        -1 + sqrt(2)
-    assert radsimp(1/(r2 + r3)) == \
-        -sqrt(2) + sqrt(3)
-    assert fraction(radsimp(1/(1 + r2 + r3))) == \
-        (-sqrt(6) + sqrt(2) + 2, 4)
-    assert fraction(radsimp(1/(r2 + r3 + r5))) == \
-        (-sqrt(30) + 2*sqrt(3) + 3*sqrt(2), 12)
-    assert fraction(radsimp(1/(1 + r2 + r3 + r5))) == (
-        (-34*sqrt(10) - 26*sqrt(15) - 55*sqrt(3) - 61*sqrt(2) + 14*sqrt(30) +
-        93 + 46*sqrt(6) + 53*sqrt(5), 71))
-    assert fraction(radsimp(1/(r2 + r3 + r5 + r7))) == (
-        (-50*sqrt(42) - 133*sqrt(5) - 34*sqrt(70) - 145*sqrt(3) + 22*sqrt(105)
-        + 185*sqrt(2) + 62*sqrt(30) + 135*sqrt(7), 215))
-    z = radsimp(1/(1 + r2/3 + r3/5 + r5 + r7))
-    assert len((3616791619821680643598*z).args) == 16
-    assert radsimp(1/z) == 1/z
-    assert radsimp(1/z, max_terms=20).expand() == 1 + r2/3 + r3/5 + r5 + r7
-    assert radsimp(1/(r2*3)) == \
-        sqrt(2)/6
-    assert radsimp(1/(r2*a + r3 + r5 + r7)) == (
-        (8*sqrt(2)*a**7 - 8*sqrt(7)*a**6 - 8*sqrt(5)*a**6 - 8*sqrt(3)*a**6 -
-        180*sqrt(2)*a**5 + 8*sqrt(30)*a**5 + 8*sqrt(42)*a**5 + 8*sqrt(70)*a**5
-        - 24*sqrt(105)*a**4 + 84*sqrt(3)*a**4 + 100*sqrt(5)*a**4 +
-        116*sqrt(7)*a**4 - 72*sqrt(70)*a**3 - 40*sqrt(42)*a**3 -
-        8*sqrt(30)*a**3 + 782*sqrt(2)*a**3 - 462*sqrt(3)*a**2 -
-        302*sqrt(7)*a**2 - 254*sqrt(5)*a**2 + 120*sqrt(105)*a**2 -
-        795*sqrt(2)*a - 62*sqrt(30)*a + 82*sqrt(42)*a + 98*sqrt(70)*a -
-        118*sqrt(105) + 59*sqrt(7) + 295*sqrt(5) + 531*sqrt(3))/(16*a**8 -
-        480*a**6 + 3128*a**4 - 6360*a**2 + 3481))
-    assert radsimp(1/(r2*a + r2*b + r3 + r7)) == (
-        (sqrt(2)*a*(a + b)**2 - 5*sqrt(2)*a + sqrt(42)*a + sqrt(2)*b*(a +
-        b)**2 - 5*sqrt(2)*b + sqrt(42)*b - sqrt(7)*(a + b)**2 - sqrt(3)*(a +
-        b)**2 - 2*sqrt(3) + 2*sqrt(7))/(2*a**4 + 8*a**3*b + 12*a**2*b**2 -
-        20*a**2 + 8*a*b**3 - 40*a*b + 2*b**4 - 20*b**2 + 8))
-    assert radsimp(1/(r2*a + r2*b + r2*c + r2*d)) == \
-        sqrt(2)/(2*a + 2*b + 2*c + 2*d)
-    assert radsimp(1/(1 + r2*a + r2*b + r2*c + r2*d)) == (
-        (sqrt(2)*a + sqrt(2)*b + sqrt(2)*c + sqrt(2)*d - 1)/(2*a**2 + 4*a*b +
-        4*a*c + 4*a*d + 2*b**2 + 4*b*c + 4*b*d + 2*c**2 + 4*c*d + 2*d**2 - 1))
-    assert radsimp((y**2 - x)/(y - sqrt(x))) == \
-        sqrt(x) + y
-    assert radsimp(-(y**2 - x)/(y - sqrt(x))) == \
-        -(sqrt(x) + y)
-    assert radsimp(1/(1 - I + a*I)) == \
-        (-I*a + 1 + I)/(a**2 - 2*a + 2)
-    assert radsimp(1/((-x + y)*(x - sqrt(y)))) == \
-        (-x - sqrt(y))/((x - y)*(x**2 - y))
-    e = (3 + 3*sqrt(2))*x*(3*x - 3*sqrt(y))
-    assert radsimp(e) == x*(3 + 3*sqrt(2))*(3*x - 3*sqrt(y))
-    assert radsimp(1/e) == (
-        (-9*x + 9*sqrt(2)*x - 9*sqrt(y) + 9*sqrt(2)*sqrt(y))/(9*x*(9*x**2 -
-        9*y)))
-    assert radsimp(1 + 1/(1 + sqrt(3))) == \
-        Mul(S.Half, -1 + sqrt(3), evaluate=False) + 1
-    A = symbols("A", commutative=False)
-    assert radsimp(x**2 + sqrt(2)*x**2 - sqrt(2)*x*A) == \
-        x**2 + sqrt(2)*x**2 - sqrt(2)*x*A
-    assert radsimp(1/sqrt(5 + 2 * sqrt(6))) == -sqrt(2) + sqrt(3)
-    assert radsimp(1/sqrt(5 + 2 * sqrt(6))**3) == -(-sqrt(3) + sqrt(2))**3
-
-    # issue 6532
-    assert fraction(radsimp(1/sqrt(x))) == (sqrt(x), x)
-    assert fraction(radsimp(1/sqrt(2*x + 3))) == (sqrt(2*x + 3), 2*x + 3)
-    assert fraction(radsimp(1/sqrt(2*(x + 3)))) == (sqrt(2*x + 6), 2*x + 6)
-
-    # issue 5994
-    e = S('-(2 + 2*sqrt(2) + 4*2**(1/4))/'
-        '(1 + 2**(3/4) + 3*2**(1/4) + 3*sqrt(2))')
-    assert radsimp(e).expand() == -2*2**(S(3)/4) - 2*2**(S(1)/4) + 2 + 2*sqrt(2)
-
-    # issue 5986 (modifications to radimp didn't initially recognize this so
-    # the test is included here)
-    assert radsimp(1/(-sqrt(5)/2 - S(1)/2 + (-sqrt(5)/2 - S(1)/2)**2)) == 1
-
-    # from issue 5934
-    eq = (
-        (-240*sqrt(2)*sqrt(sqrt(5) + 5)*sqrt(8*sqrt(5) + 40) -
-        360*sqrt(2)*sqrt(-8*sqrt(5) + 40)*sqrt(-sqrt(5) + 5) -
-        120*sqrt(10)*sqrt(-8*sqrt(5) + 40)*sqrt(-sqrt(5) + 5) +
-        120*sqrt(2)*sqrt(-sqrt(5) + 5)*sqrt(8*sqrt(5) + 40) +
-        120*sqrt(2)*sqrt(-8*sqrt(5) + 40)*sqrt(sqrt(5) + 5) +
-        120*sqrt(10)*sqrt(-sqrt(5) + 5)*sqrt(8*sqrt(5) + 40) +
-        120*sqrt(10)*sqrt(-8*sqrt(5) + 40)*sqrt(sqrt(5) + 5))/(-36000 -
-        7200*sqrt(5) + (12*sqrt(10)*sqrt(sqrt(5) + 5) +
-        24*sqrt(10)*sqrt(-sqrt(5) + 5))**2))
-    assert radsimp(eq) is S.NaN  # it's 0/0
-
-    # work with normal form
-    e = 1/sqrt(sqrt(7)/7 + 2*sqrt(2) + 3*sqrt(3) + 5*sqrt(5)) + 3
-    assert radsimp(e) == (
-        -sqrt(sqrt(7) + 14*sqrt(2) + 21*sqrt(3) +
-        35*sqrt(5))*(-11654899*sqrt(35) - 1577436*sqrt(210) - 1278438*sqrt(15)
-        - 1346996*sqrt(10) + 1635060*sqrt(6) + 5709765 + 7539830*sqrt(14) +
-        8291415*sqrt(21))/1300423175 + 3)
-
-    # obey power rules
-    base = sqrt(3) - sqrt(2)
-    assert radsimp(1/base**3) == (sqrt(3) + sqrt(2))**3
-    assert radsimp(1/(-base)**3) == -(sqrt(2) + sqrt(3))**3
-    assert radsimp(1/(-base)**x) == (-base)**(-x)
-    assert radsimp(1/base**x) == (sqrt(2) + sqrt(3))**x
-    assert radsimp(root(1/(-1 - sqrt(2)), -x)) == (-1)**(-1/x)*(1 + sqrt(2))**(1/x)
-
-    # recurse
-    e = cos(1/(1 + sqrt(2)))
-    assert radsimp(e) == cos(-sqrt(2) + 1)
-    assert radsimp(e/2) == cos(-sqrt(2) + 1)/2
-    assert radsimp(1/e) == 1/cos(-sqrt(2) + 1)
-    assert radsimp(2/e) == 2/cos(-sqrt(2) + 1)
-    assert fraction(radsimp(e/sqrt(x))) == (sqrt(x)*cos(-sqrt(2)+1), x)
-
-    # test that symbolic denominators are not processed
-    r = 1 + sqrt(2)
-    assert radsimp(x/r, symbolic=False) == -x*(-sqrt(2) + 1)
-    assert radsimp(x/(y + r), symbolic=False) == x/(y + 1 + sqrt(2))
-    assert radsimp(x/(y + r)/r, symbolic=False) == \
-        -x*(-sqrt(2) + 1)/(y + 1 + sqrt(2))
-
-    # issue 7408
-    eq = sqrt(x)/sqrt(y)
-    assert radsimp(eq) == umul(sqrt(x), sqrt(y), 1/y)
-    assert radsimp(eq, symbolic=False) == eq
-
-    # issue 7498
-    assert radsimp(sqrt(x)/sqrt(y)**3) == umul(sqrt(x), sqrt(y**3), 1/y**3)
-
-    # for coverage
-    eq = sqrt(x)/y**2
-    assert radsimp(eq) == eq
-
-
-def test_radsimp_issue_3214():
-    c, p = symbols('c p', positive=True)
-    s = sqrt(c**2 - p**2)
-    b = (c + I*p - s)/(c + I*p + s)
-    assert radsimp(b) == -I*(c + I*p - sqrt(c**2 - p**2))**2/(2*c*p)
-
-
-def test_collect_const():
-    # coverage not provided by above tests
-    assert collect_const(2*sqrt(3) + 4*a*sqrt(5)) == \
-        2*(2*sqrt(5)*a + sqrt(3))  # let the primitive reabsorb
-    assert collect_const(2*sqrt(3) + 4*a*sqrt(5), sqrt(3)) == \
-        2*sqrt(3) + 4*a*sqrt(5)
-    assert collect_const(sqrt(2)*(1 + sqrt(2)) + sqrt(3) + x*sqrt(2)) == \
-        sqrt(2)*(x + 1 + sqrt(2)) + sqrt(3)
-
-    # issue 5290
-    assert collect_const(2*x + 2*y + 1, 2) == \
-        collect_const(2*x + 2*y + 1) == \
-        Add(S(1), Mul(2, x + y, evaluate=False), evaluate=False)
-    assert collect_const(-y - z) == Mul(-1, y + z, evaluate=False)
-    assert collect_const(2*x - 2*y - 2*z, 2) == \
-        Mul(2, x - y - z, evaluate=False)
-    assert collect_const(2*x - 2*y - 2*z, -2) == \
-        _unevaluated_Add(2*x, Mul(-2, y + z, evaluate=False))
-
-    # this is why the content_primitive is used
-    eq = (sqrt(15 + 5*sqrt(2))*x + sqrt(3 + sqrt(2))*y)*2
-    assert collect_sqrt(eq + 2) == \
-        2*sqrt(sqrt(2) + 3)*(sqrt(5)*x + y) + 2
-
-
-def test_issue_5933():
-    from sympy import Polygon, RegularPolygon, denom
-    x = Polygon(*RegularPolygon((0, 0), 1, 5).vertices).centroid.x
-    assert abs(denom(x).n()) > 1e-12
-    assert abs(denom(radsimp(x))) > 1e-12  # in case simplify didn't handle it
-
-
-def test_fraction_expand():
-    eq = (x + y)*y/x
-    assert eq.expand(frac=True) == fraction_expand(eq) == (x*y + y**2)/x
-    assert eq.expand() == y + y**2/x
-
-
-def test_combsimp_gamma():
-    from sympy.abc import x, y
-    R = Rational
-
-    assert combsimp(gamma(x)) == gamma(x)
-    assert combsimp(gamma(x + 1)/x) == gamma(x)
-    assert combsimp(gamma(x)/(x - 1)) == gamma(x - 1)
-    assert combsimp(x*gamma(x)) == gamma(x + 1)
-    assert combsimp((x + 1)*gamma(x + 1)) == gamma(x + 2)
-    assert combsimp(gamma(x + y)*(x + y)) == gamma(x + y + 1)
-    assert combsimp(x/gamma(x + 1)) == 1/gamma(x)
-    assert combsimp((x + 1)**2/gamma(x + 2)) == (x + 1)/gamma(x + 1)
-    assert combsimp(x*gamma(x) + gamma(x + 3)/(x + 2)) == \
-        (x + 2)*gamma(x + 1)
-
-    assert combsimp(gamma(2*x)*x) == gamma(2*x + 1)/2
-    assert combsimp(gamma(2*x)/(x - S(1)/2)) == 2*gamma(2*x - 1)
-
-    assert combsimp(gamma(x)*gamma(1 - x)) == pi/sin(pi*x)
-    assert combsimp(gamma(x)*gamma(-x)) == -pi/(x*sin(pi*x))
-    assert combsimp(1/gamma(x + 3)/gamma(1 - x)) == \
-        sin(pi*x)/(pi*x*(x + 1)*(x + 2))
-
-    assert powsimp(combsimp(
-        gamma(x)*gamma(x + S(1)/2)*gamma(y)/gamma(x + y))) == \
-        2**(-2*x + 1)*sqrt(pi)*gamma(2*x)*gamma(y)/gamma(x + y)
-    assert combsimp(1/gamma(x)/gamma(x - S(1)/3)/gamma(x + S(1)/3)) == \
-        3**(3*x - S(3)/2)/(2*pi*gamma(3*x - 1))
-    assert simplify(
-        gamma(S(1)/2 + x/2)*gamma(1 + x/2)/gamma(1 + x)/sqrt(pi)*2**x) == 1
-    assert combsimp(gamma(S(-1)/4)*gamma(S(-3)/4)) == 16*sqrt(2)*pi/3
-
-    assert powsimp(combsimp(gamma(2*x)/gamma(x))) == \
-        2**(2*x - 1)*gamma(x + S(1)/2)/sqrt(pi)
-
-    # issue 6792
-    e = (-gamma(k)*gamma(k + 2) + gamma(k + 1)**2)/gamma(k)**2
-    assert combsimp(e) == -k
-    assert combsimp(1/e) == -1/k
-    e = (gamma(x) + gamma(x + 1))/gamma(x)
-    assert combsimp(e) == x + 1
-    assert combsimp(1/e) == 1/(x + 1)
-    e = (gamma(x) + gamma(x + 2))*(gamma(x - 1) + gamma(x))/gamma(x)
-    assert combsimp(e) == (x**2 + x + 1)*gamma(x + 1)/(x - 1)
-    e = (-gamma(k)*gamma(k + 2) + gamma(k + 1)**2)/gamma(k)**2
-    assert combsimp(e**2) == k**2
-    assert combsimp(e**2/gamma(k + 1)) == k/gamma(k)
-    a = R(1, 2) + R(1, 3)
-    b = a + R(1, 3)
-    assert combsimp(gamma(2*k)/gamma(k)*gamma(k + a)*gamma(k + b))
-    3*2**(2*k + 1)*3**(-3*k - 2)*sqrt(pi)*gamma(3*k + R(3, 2))/2
-
-    A, B = symbols('A B', commutative=False)
-    assert combsimp(e*B*A) == combsimp(e)*B*A
-
-    # check iteration
-    assert combsimp(gamma(2*k)/gamma(k)*gamma(-k - R(1, 2))) == (
-        -2**(2*k + 1)*sqrt(pi)/(2*((2*k + 1)*cos(pi*k))))
-    assert combsimp(
-        gamma(k)*gamma(k + R(1, 3))*gamma(k + R(2, 3))/gamma(3*k/2)) == (
-        3*2**(3*k + 1)*3**(-3*k - S.Half)*sqrt(pi)*gamma(3*k/2 + S.Half)/2)
-
-
-def test_polarify():
-    from sympy import polar_lift, polarify
-    x = Symbol('x')
-    z = Symbol('z', polar=True)
-    f = Function('f')
-    ES = {}
-
-    assert polarify(-1) == (polar_lift(-1), ES)
-    assert polarify(1 + I) == (polar_lift(1 + I), ES)
-
-    assert polarify(exp(x), subs=False) == exp(x)
-    assert polarify(1 + x, subs=False) == 1 + x
-    assert polarify(f(I) + x, subs=False) == f(polar_lift(I)) + x
-
-    assert polarify(x, lift=True) == polar_lift(x)
-    assert polarify(z, lift=True) == z
-    assert polarify(f(x), lift=True) == f(polar_lift(x))
-    assert polarify(1 + x, lift=True) == polar_lift(1 + x)
-    assert polarify(1 + f(x), lift=True) == polar_lift(1 + f(polar_lift(x)))
-
-    newex, subs = polarify(f(x) + z)
-    assert newex.subs(subs) == f(x) + z
-
-    mu = Symbol("mu")
-    sigma = Symbol("sigma", positive=True)
-
-    # Make sure polarify(lift=True) doesn't try to lift the integration
-    # variable
-    assert polarify(
-        Integral(sqrt(2)*x*exp(-(-mu + x)**2/(2*sigma**2))/(2*sqrt(pi)*sigma),
-        (x, -oo, oo)), lift=True) == Integral(sqrt(2)*(sigma*exp_polar(0))**exp_polar(I*pi)*
-        exp((sigma*exp_polar(0))**(2*exp_polar(I*pi))*exp_polar(I*pi)*polar_lift(-mu + x)**
-        (2*exp_polar(0))/2)*exp_polar(0)*polar_lift(x)/(2*sqrt(pi)), (x, -oo, oo))
-
-
-def test_unpolarify():
-    from sympy import (exp_polar, polar_lift, exp, unpolarify,
-                       principal_branch)
-    from sympy import gamma, erf, sin, tanh, uppergamma, Eq, Ne
-    from sympy.abc import x
-    p = exp_polar(7*I) + 1
-    u = exp(7*I) + 1
-
-    assert unpolarify(1) == 1
-    assert unpolarify(p) == u
-    assert unpolarify(p**2) == u**2
-    assert unpolarify(p**x) == p**x
-    assert unpolarify(p*x) == u*x
-    assert unpolarify(p + x) == u + x
-    assert unpolarify(sqrt(sin(p))) == sqrt(sin(u))
-
-    # Test reduction to principal branch 2*pi.
-    t = principal_branch(x, 2*pi)
-    assert unpolarify(t) == x
-    assert unpolarify(sqrt(t)) == sqrt(t)
-
-    # Test exponents_only.
-    assert unpolarify(p**p, exponents_only=True) == p**u
-    assert unpolarify(uppergamma(x, p**p)) == uppergamma(x, p**u)
-
-    # Test functions.
-    assert unpolarify(sin(p)) == sin(u)
-    assert unpolarify(tanh(p)) == tanh(u)
-    assert unpolarify(gamma(p)) == gamma(u)
-    assert unpolarify(erf(p)) == erf(u)
-    assert unpolarify(uppergamma(x, p)) == uppergamma(x, p)
-
-    assert unpolarify(uppergamma(sin(p), sin(p + exp_polar(0)))) == \
-        uppergamma(sin(u), sin(u + 1))
-    assert unpolarify(uppergamma(polar_lift(0), 2*exp_polar(0))) == \
-        uppergamma(0, 2)
-
-    assert unpolarify(Eq(p, 0)) == Eq(u, 0)
-    assert unpolarify(Ne(p, 0)) == Ne(u, 0)
-    assert unpolarify(polar_lift(x) > 0) == (x > 0)
-
-    # Test bools
-    assert unpolarify(True) is True
-
-
-def test_issue_6097():
-    assert collect(a*y**(2.0*x) + b*y**(2.0*x), y**x) == y**(2.0*x)*(a + b)
-    assert collect(a*2**(2.0*x) + b*2**(2.0*x), 2**x) == 2**(2.0*x)*(a + b)
-
-
 def test_signsimp():
     e = x*(-x + 1) + x*(x - 1)
     assert signsimp(Eq(e, 0)) is S.true
@@ -1788,10 +533,6 @@ def test_Piecewise():
     assert simplify(Piecewise((e1, x < e2), (e3, True))) == \
         Piecewise((s1, x < s2), (s3, True))
 
-    # trigsimp tries not to touch non-trig containing args
-    assert trigsimp(Piecewise((e1, e3 < e2), (e3, True))) == \
-        Piecewise((e1, e3 < s2), (e3, True))
-
 
 def test_polymorphism():
     class A(Basic):
@@ -1805,11 +546,6 @@ def test_polymorphism():
 def test_issue_from_PR1599():
     n1, n2, n3, n4 = symbols('n1 n2 n3 n4', negative=True)
     assert simplify(I*sqrt(n1)) == -sqrt(-n1)
-    assert (powsimp(sqrt(n1)*sqrt(n2)*sqrt(n3)) ==
-        -I*sqrt(-n1)*sqrt(-n2)*sqrt(-n3))
-    assert (powsimp(root(n1, 3)*root(n2, 3)*root(n3, 3)*root(n4, 3)) ==
-        -(-1)**(S(1)/3)*
-        (-n1)**(S(1)/3)*(-n2)**(S(1)/3)*(-n3)**(S(1)/3)*(-n4)**(S(1)/3))
 
 
 def test_issue_6811():
@@ -1818,13 +554,6 @@ def test_issue_6811():
     # reject the 2-arg Mul -- these are a headache for test writing
     assert simplify(eq.expand()) == \
         2*x**2 + 4*x*y + 2*x + 4*y
-
-
-@XFAIL
-def test_issue_6811_fail():
-    xp, y, x, z = symbols('xp, y, x, z')
-    eq = 4*(-19*sin(x)*y + 5*sin(3*x)*y + 15*cos(2*x)*z - 21*z)*xp/(9*cos(x) - 5*cos(3*x))
-    assert trigsimp(eq) == -2*(2*cos(x)*tan(x)*y + 3*z)*xp/cos(x)
 
 
 def test_issue_6920():
@@ -1844,78 +573,6 @@ def test_issue_7001():
         Piecewise((-1, r <= R), (0, True))
 
 
-def test_exptrigsimp():
-    def valid(a, b):
-        from sympy.utilities.randtest import verify_numerically as tn
-        if not (tn(a, b) and a == b):
-            return False
-        return True
-
-    assert exptrigsimp(exp(x) + exp(-x)) == 2*cosh(x)
-    assert exptrigsimp(exp(x) - exp(-x)) == 2*sinh(x)
-    e = [cos(x) + I*sin(x), cos(x) - I*sin(x),
-         cosh(x) - sinh(x), cosh(x) + sinh(x)]
-    ok = [exp(I*x), exp(-I*x), exp(-x), exp(x)]
-    assert all(valid(i, j) for i, j in zip(
-        [exptrigsimp(ei) for ei in e], ok))
-
-    ue = [cos(x) + sin(x), cos(x) - sin(x),
-          cosh(x) + I*sinh(x), cosh(x) - I*sinh(x)]
-    assert [exptrigsimp(ei) == ei for ei in ue]
-
-    res = []
-    ok = [y*tanh(1), 1/(y*tanh(1)), I*y*tan(1), -I/(y*tan(1)),
-        y*tanh(x), 1/(y*tanh(x)), I*y*tan(x), -I/(y*tan(x)),
-        y*tanh(1 + I), 1/(y*tanh(1 + I))]
-    for a in (1, I, x, I*x, 1 + I):
-        w = exp(a)
-        eq = y*(w - 1/w)/(w + 1/w)
-        s = simplify(eq)
-        assert s == exptrigsimp(eq)
-        res.append(s)
-        sinv = simplify(1/eq)
-        assert sinv == exptrigsimp(1/eq)
-        res.append(sinv)
-    assert all(valid(i, j) for i, j in zip(res, ok))
-
-    for a in range(1, 3):
-        w = exp(a)
-        e = w + 1/w
-        s = simplify(e)
-        assert s == exptrigsimp(e)
-        assert valid(s, 2*cosh(a))
-        e = w - 1/w
-        s = simplify(e)
-        assert s == exptrigsimp(e)
-        assert valid(s, 2*sinh(a))
-
-
-def test_issue_2827_trigsimp_methods():
-    def measure1(expr):
-        return len(str(expr))
-
-    def measure2(expr):
-        return -count_ops(expr)
-                                       # Return the most complicated result
-    expr = (x + 1)/(x + sin(x)**2 + cos(x)**2)
-    ans = Matrix([1])
-    M = Matrix([expr])
-    assert trigsimp(M, method='fu', measure=measure1) == ans
-    assert trigsimp(M, method='fu', measure=measure2) != ans
-    # all methods should work with Basic expressions even if they
-    # aren't Expr
-    M = Matrix.eye(1)
-    assert all(trigsimp(M, method=m) == M for m in
-        'fu matching groebner old'.split())
-    # watch for E in exptrigsimp, not only exp()
-    eq = 1/sqrt(E) + E
-    assert exptrigsimp(eq) == eq
-
-
-def test_powsimp_on_numbers():
-    assert 2**(S(1)/3 - 2) == 2**(S(1)/3)/4
-
-
 def test_inequality_no_auto_simplify():
     # no simplify on creation but can be simplified
     lhs = cos(x)**2 + sin(x)**2
@@ -1923,19 +580,6 @@ def test_inequality_no_auto_simplify():
     e = Lt(lhs, rhs)
     assert e == Lt(lhs, rhs, evaluate=False)
     assert simplify(e)
-
-
-def test_issue_9324_simplify():
-    M = MatrixSymbol('M', 10, 10)
-    e = M[0, 0] + M[5, 4] + 1304
-    assert simplify(e) == e
-
-
-def test_issue_9324_powsimp_on_matrix_symbol():
-    M = MatrixSymbol('M', 10, 10)
-    expr = powsimp(M, deep=True)
-    assert expr == M
-    assert expr.args[0] == 'M'
 
 
 def test_issue_9398():
